requirements.txt
flask
torch
ultralytics
Pillow
